// src/api/models/login-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ling Workshop API
 * API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UserInfoDto } from './user-info-dto';

/**
 * 
 * @export
 * @interface LoginResponseDto
 */
export interface LoginResponseDto {
    /**
     * 响应消息
     * @type {string}
     * @memberof LoginResponseDto
     */
    'message': string;
    /**
     * 用户信息
     * @type {UserInfoDto}
     * @memberof LoginResponseDto
     */
    'user': UserInfoDto;
    /**
     * 访问令牌
     * @type {string}
     * @memberof LoginResponseDto
     */
    'accessToken': string;
    /**
     * 刷新令牌
     * @type {string}
     * @memberof LoginResponseDto
     */
    'refreshToken': string;
}