// src/api/models/user-info-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ling Workshop API
 * API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UserInfoDto
 */
export interface UserInfoDto {
    /**
     * 用户ID
     * @type {object}
     * @memberof UserInfoDto
     */
    'id': object;
    /**
     * 邮箱
     * @type {string}
     * @memberof UserInfoDto
     */
    'email': string;
    /**
     * 用户名
     * @type {string}
     * @memberof UserInfoDto
     */
    'name': string;
    /**
     * 头像URL
     * @type {object}
     * @memberof UserInfoDto
     */
    'avatarUrl': object | null;
    /**
     * 邮箱是否已验证
     * @type {boolean}
     * @memberof UserInfoDto
     */
    'isEmailVerified': boolean;
    /**
     * 用户状态
     * @type {string}
     * @memberof UserInfoDto
     */
    'status'?: string;
}